requirements.txt
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
gitdb==4.0.12
GitPython==3.1.45
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.13.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.5.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
SQLAlchemy==2.0.45
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.15
